Move validating chain config
